stock-alerts/requirements.txt
azure-functions
yfinance
pandas
